[modules/archive] avoid overwriting an existing thumbnail

+++ ipol_demo/modules/archive/archive.py
@@ -299,8 +299,9 @@ class Archive():
         """
         try:
             final_path, _ = self.get_new_path(main_dir, hash_file, extension)
-            shutil.copyfile(original_path, final_path)
-            return final_path
+            if not os.path.exists(final_path):
+                shutil.copyfile(original_path, final_path)
+                return final_path
         except Exception as ex:
             message = "Failure in copy_file_in_folder. Error={}".format(ex)
             self.logger.exception(message)
